XtreamSplit: -refactored -checkComplete processes information from xtm header (number of Parts, missing Parts) -updated source (eg added comments for archive headers) -updated md5 check -updated final filesize calculation (my example xtm had buggy headers) UnixSplit: -refactored -updated checkComplete ExtractionExtension: -major refactoring -updated to use new FileBytesCacheFlusher -updated skip handling -added isFileSupported for Dialogs(faster than isArchivSupported) -added many new supported formats -optimized for less ressource/higher performance -updated to use existing instance of new instance (where needed)
HachaSplit:
-fixes #999
HJSplit:
-updated cutKiller handling
-updated checkComplete
SplitType:
-new enum to provide all information needed for merging supported split formats
ArchiveType:
-new enum to provide all information needed for extracting supported archive formats
Multi:
-updated to use new ArchiveType, reduces code and complexity
-updated arm library handling
-updated checkComplete
-extract now uses Seven7ExtractCallback for bzip2
-added null (Long) Size support (eg Tar,Bzip2)
-updated skip handling
-cleaner support for different archive formats
MultiCallback:
-updated to use FileBytesCache
MultiOpener/RarOpener:
-fixed some issues that occured with different archive formats (eg tar,bzip2)
DummyArchiveDialog:
-shows archiveFormat type
FileArchive:
-optimized for major speed up
-uses nio when available
-updated for less IO overhead -> speed up
DownloadLinkArchive:
-optimized for major speed up
-fixed isComplete
CrawledLinkFactory:
-optimized for major speed up
ExtractAction:
-updated for new isFileSupported
-fixed possible npe
Archive:
-added skippedFiles (to have reference to skipped files) and differ between extracted/skipped files
DummyArchive/File:
-updated for speed up
ExtractionListenerList:
-updated cleanup stuff
Other:
-fixes #63324
-fixes #8746





git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@29227 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/org/jdownloader/extensions/extraction/ExtractionListenerList.java
@@ -36,6 +36,7 @@ import jd.controlling.TaskQueue;
 import jd.controlling.packagecontroller.AbstractNode;
 import jd.plugins.DownloadLink;
 import jd.plugins.FilePackage;
+import net.sf.sevenzipjbinding.ArchiveFormat;
 
 import org.appwork.swing.MigPanel;
 import org.appwork.uio.UIOManager;
@@ -358,10 +359,13 @@ public class ExtractionListenerList implements ExtractionListener {
 
                     @Override
                     protected Void run() throws RuntimeException {
-                        FileCreationManager.getInstance().getEventSender().fireEvent(new FileCreationEvent(controller, FileCreationEvent.Type.NEW_FILES, controller.getArchiv().getExtractedFiles().toArray(new File[controller.getArchiv().getExtractedFiles().size()])));
-                        if (ex.getSettings().isDeleteInfoFilesAfterExtraction()) {
-                            File fileOutput = new File(controller.getArchiv().getFirstArchiveFile().getFilePath());
-                            File infoFiles = new File(fileOutput.getParentFile(), fileOutput.getName().replaceFirst("(?i)(\\.pa?r?t?\\.?[0-9]+\\.rar|\\.rar)$", "") + ".info");
+                        final Archive archive = controller.getArchiv();
+                        final ArrayList<File> files = new ArrayList<File>(archive.getExtractedFiles());
+                        files.addAll(archive.getSkippedFiles());
+                        FileCreationManager.getInstance().getEventSender().fireEvent(new FileCreationEvent(controller, FileCreationEvent.Type.NEW_FILES, files.toArray(new File[files.size()])));
+                        if (ex.getSettings().isDeleteInfoFilesAfterExtraction() && ArchiveFormat.RAR.equals(archive.getType().getArchiveFormat())) {
+                            final File fileOutput = new File(archive.getFirstArchiveFile().getFilePath());
+                            final File infoFiles = new File(fileOutput.getParentFile(), fileOutput.getName().replaceFirst("(?i)(\\.pa?r?t?\\.?[0-9]+\\.rar|\\.rar)$", "") + ".info");
                             if (infoFiles.exists() && infoFiles.delete()) {
                                 logger.info(infoFiles.getName() + " removed");
                             }
@@ -390,15 +394,7 @@ public class ExtractionListenerList implements ExtractionListener {
                 }
                 if (controller.gotKilled()) {
                     controller.getArchiv().getFirstArchiveFile().setMessage(controller, null);
-                    for (File f : controller.getArchiv().getExtractedFiles()) {
-                        if (f.exists()) {
-                            if (!FileCreationManager.getInstance().delete(f, null)) {
-                                logger.warning("Could not delete file " + f.getAbsolutePath());
-                            } else {
-                                logger.warning("Deleted file " + f.getAbsolutePath());
-                            }
-                        }
-                    }
+                    cleanUpIncomplete(controller);
                 }
             } finally {
                 controller.getArchiv().setActive(false);
@@ -429,15 +425,7 @@ public class ExtractionListenerList implements ExtractionListener {
                 } else {
                     controller.getArchiv().setStatus(controller, ExtractionStatus.ERRROR_FILE_NOT_FOUND);
                 }
-                for (File f : controller.getArchiv().getExtractedFiles()) {
-                    if (f.exists()) {
-                        if (!FileCreationManager.getInstance().delete(f, null)) {
-                            logger.warning("Could not delete file " + f.getAbsolutePath());
-                        } else {
-                            logger.warning("Deleted file " + f.getAbsolutePath());
-                        }
-                    }
-                }
+                cleanUpIncomplete(controller);
             } finally {
                 controller.getArchiv().setActive(false);
                 ex.onFinished(controller);
